Improvements to the splitting code (#457)
1. Separated functions.py into modules
2. Replaced the ExecutorFactory class by the method in StreamFunction
